git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/lucene5376@1573003 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/codecs/src/java/org/apache/lucene/codecs/simpletext/SimpleTextFieldsReader.java
@@ -562,7 +562,7 @@ class SimpleTextFieldsReader extends FieldsProducer {
           termCount++;
         }
       }
-      docCount = (int) visitedDocs.cardinality();
+      docCount = visitedDocs.cardinality();
       fst = b.finish();
       /*
       PrintStream ps = new PrintStream("out.dot");
